Move code to designated packages
