+++ json-streaming/src/test/java/docs/javadsl/JsonReaderUsageTest.java
@@ -2,7 +2,7 @@
  * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.stream.alpakka.json;
+package docs.javadsl;
 
 import akka.actor.ActorSystem;
 import akka.stream.ActorMaterializer;
